app/requirements.txt
pandas
numpy
requests
pyarrow
fastparquet
matplotlib
scikit-learn
xgboost
holidays
psycopg2-binary
pyyaml
shap
streamlit
fastapi
uvicorn